+++ R/io.R
@@ -55,12 +55,36 @@ invalid_utf8 = function(x) {
 #' @examples
 #' xfun::file_string(system.file('DESCRIPTION', package = 'xfun'))
 file_string = function(file) {
-  raw_string(paste(read_utf8(file), collapse = '\n'))
+  x = read_utf8(file)
+  # paste converts 0-length character() into 1-length ""
+  if (length(x)) x = paste(x, collapse = '\n')
+  raw_string(x)
 }
 
 # read all records of a binary file as a raw vector by default
 read_bin = function(file, what = 'raw', n = file.info(file)$size, ...) {
   readBin(file, what, n, ...)
+}
+
+#' Read a text file, process the text with a function, and write the text back
+#'
+#' Read a text file with the UTF-8 encoding, apply a function to the text, and
+#' write back to the original file.
+#' @param file Path to a text file.
+#' @param FUN A function to process the text.
+#' @param x The content of the file.
+#' @return If \code{file} is provided, invisible \code{NULL} (the file is
+#'   updated as a side effect), otherwise the processed content (as a character
+#'   vector).
+#' @export
+#' @examples f = tempfile()
+#' xfun::write_utf8('Hello World', f)
+#' xfun::process_file(f, function(x) gsub('World', 'woRld', x))
+#' xfun::read_utf8(f)  # see if it has been updated
+#' file.remove(f)
+process_file = function(file, FUN = identity, x = read_utf8(file)) {
+  x = FUN(x)
+  if (missing(file)) x else write_utf8(x, file)
 }
 
 #' Search and replace strings in files
@@ -120,4 +144,81 @@ gsub_dir = function(..., dir = '.', recursive = TRUE, ext = NULL, mimetype = '.*
 #' @export
 gsub_ext = function(ext, ..., dir = '.', recursive = TRUE) {
   gsub_dir(..., dir = dir, recursive = recursive, ext = ext)
+}
+
+#' Perform replacement with \code{gsub()} on elements matched from \code{grep()}
+#'
+#' This function is a shorthand of \code{gsub(pattern, replacement,
+#' grep(pattern, x, value = TRUE))}.
+#' @param pattern,replacement,x,... Passed to \code{\link{grep}()} and
+#'   \code{gsub()}.
+#' @return A character vector.
+#' @export
+#' @examples # find elements that matches 'a[b]+c' and capitalize 'b' with perl regex
+#' xfun::grep_sub('a([b]+)c', 'a\\U\\1c', c('abc', 'abbbc', 'addc', '123'), perl = TRUE)
+grep_sub = function(pattern, replacement, x, ...) {
+  x = grep(pattern, x, value = TRUE, ...)
+  gsub(pattern, replacement, x, ...)
+}
+
+#' Try various methods to download a file
+#'
+#' Try all possible methods in \code{\link{download.file}()} (e.g.,
+#' \code{libcurl}, \code{curl}, \code{wget}, and \code{wininet}) and see if any
+#' method can succeed. The reason to enumerate all methods is that sometimes the
+#' default method does not work, e.g.,
+#' \url{https://stat.ethz.ch/pipermail/r-devel/2016-June/072852.html}.
+#' @param url The URL of the file.
+#' @param output Path to the output file. By default, it is determined by
+#'   \code{\link{url_filename}()}.
+#' @param ... Other arguments to be passed to \code{\link{download.file}()}
+#'   (except \code{method}).
+#' @return The integer code \code{0} for success, or an error if none of the
+#'   methods work.
+#' @export
+download_file = function(url, output = url_filename(url), ...) {
+  download = function(method = 'auto') download.file(url, output, ..., method = method)
+  for (method in c(if (is_windows()) 'wininet', 'libcurl', 'auto')) {
+    if (!inherits(try_silent(res <- download(method = method)), 'try-error') && res == 0)
+      return(res)
+  }
+
+  # check for libcurl/curl/wget/lynx, call download.file with appropriate method
+  res = NA
+  if (Sys.which('curl') != '') {
+    # curl needs to add a -L option to follow redirects
+    opts = if (is.null(getOption('download.file.extra'))) options(download.file.extra = '-L')
+    res = download(method = 'curl'); options(opts)
+    if (res == 0) return(res)
+  }
+  if (Sys.which('wget') != '') {
+    if ((res <- download(method = 'wget')) == 0) return(res)
+  }
+  if (Sys.which('lynx') != '') {
+    if ((res <- download(method = 'lynx')) == 0) return(res)
+  }
+  if (is.na(res)) stop('No download method works (auto/wininet/wget/curl/lynx)')
+
+  res
+}
+
+#' Get the tags of Github releases of a repository
+#'
+#' Read the HTML source of the release page and parse the tags of the releases.
+#' @param repo The repository name of the form \code{user/repo}, e.g.,
+#'   \code{"yihui/xfun"}.
+#' @param subpath A character string to be appended to the URL of Github
+#'   releases (i.e., \verb{https://github.com/user/repo/releases/}). For
+#'   example, you may use \code{subpath = "latest"} to get the tag of the latest
+#'   release.
+#' @param pattern A regular expression to extract the tags from the HTML source.
+#'   It must contain a group (i.e., must have a pair of parentheses).
+#' @export
+#' @return A character vector of (GIT) tags.
+#' @examples if (interactive()) xfun::github_releases('yihui/xfun')
+github_releases = function(repo, subpath = '', pattern = '(v[0-9.]+)') {
+  h = readLines(sprintf('https://github.com/%s/releases/%s', repo, subpath), warn = FALSE)
+  r = sprintf('^.*?releases/tag/%s".*', pattern)
+  v = gsub(r, '\\1', grep(r, h, value = TRUE))
+  unique(v)
 }