+++ tests/testit/test-paths.R
@@ -1,6 +1,13 @@
 library(testit)
 
-assert('with_ext works for corner cases', {
+assert('file_ext() and sans_ext() work', {
+  p = c('abc.doc', 'def123.tex#', 'path/to/foo.Rmd', 'backup.ppt~', 'pkg.tar.xz')
+  (file_ext(p) %==% c('doc', 'tex#', 'Rmd', 'ppt~', 'tar.xz'))
+  (sans_ext(p) %==% c('abc', 'def123', 'path/to/foo', 'backup', 'pkg'))
+  (file_ext(c('foo.bar.gz', 'foo', 'file.nb.html')) %==% c('gz', '', 'nb.html'))
+})
+
+assert('with_ext() works for corner cases', {
   (with_ext(character(), 'abc') %==% character())
   (with_ext('abc', character()) %==% 'abc')
   (with_ext(NA_character_, 'abc') %==% NA_character_)
@@ -12,9 +19,41 @@ assert('with_ext works for corner cases', {
   (with_ext(c('a', 'b', 'c'), c('', '.d', 'e.e')) %==% c('a', 'b.d', 'c.e.e'))
 })
 
-assert('same_path works', {
+assert('same_path() works', {
   (is.na(same_path('~/foo', NA_character_)))
   (is.na(same_path(NA_character_, '~/foo')))
   (same_path('~/foo', file.path(Sys.getenv('HOME'), 'foo')))
   (!same_path(tempdir(), 'foo'))
 })
+
+assert('url_filename() returns the file names in URLs', {
+  (url_filename('https://yihui.org/images/logo.png') %==% 'logo.png')
+  (url_filename(c(
+    'https://yihui.org/index.html',
+    'https://yihui.org/index.html?foo=bar',
+    'https://yihui.org/index.html#about'
+  )) %==% rep('index.html', 3))
+})
+
+assert('is_abs_path() recognizes absolute paths on Windows and *nix', {
+  (!is_abs_path('abc/def'))
+  (is_abs_path(if (.Platform$OS.type == 'windows') {
+    c('D:\\abc', '\\\\netdrive\\somewhere')
+  } else '/abc/def'))
+})
+
+assert('del_empty_dir() correctly deletes empty dirs', {
+  # do nothing is NULL
+  (del_empty_dir(NULL) %==% NULL)
+  # remove if empty
+  dir.create(temp_dir <- tempfile())
+  del_empty_dir(temp_dir)
+  (!dir_exists(temp_dir))
+  # do not remove if not empty
+  dir.create(temp_dir <- tempfile())
+  writeLines('test', tempfile(tmpdir = temp_dir))
+  (del_empty_dir(temp_dir) %==% NULL)
+  (dir_exists(temp_dir))
+  unlink(temp_dir, recursive = TRUE)
+})
+